refactor(@ngtools/webpack): move interfaces to separate file
This is to avoid circular imports
